# reduce allocs in getSlotQueueKey